pyproject.toml
[tool.poetry]
name = "alarmbot"
version = "0.1.0"
description = ""
authors = ["Nguyễn Hồng Quân <ng.hong.quan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
aiogram = "^2.2"
gino = "^0.8.3"
sanic = "^19.6"
tortoise-orm = "^0.12.5"
dynaconf = "^2.0"
logbook = "^1.4"
babel = "^2.7"
aiohttp = "^3.5"
webargs = "^5.4"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
